Add one more test and move tests to proper place
