+++ src/test/groovy/org/prebid/server/functional/tests/BidderParamsSpec.groovy
@@ -11,7 +11,6 @@ import org.prebid.server.functional.model.request.auction.Device
 import org.prebid.server.functional.model.request.auction.Geo
 import org.prebid.server.functional.model.request.auction.Imp
 import org.prebid.server.functional.model.request.auction.Native
-import org.prebid.server.functional.model.request.auction.Prebid
 import org.prebid.server.functional.model.request.auction.PrebidStoredRequest
 import org.prebid.server.functional.model.request.auction.RegsExt
 import org.prebid.server.functional.model.request.auction.Site
@@ -23,7 +22,6 @@ import org.prebid.server.functional.model.response.auction.ErrorType
 import org.prebid.server.functional.util.PBSUtils
 import org.prebid.server.functional.util.privacy.CcpaConsent
 
-import static org.mockserver.model.HttpStatusCode.BAD_REQUEST_400
 import static org.prebid.server.functional.model.bidder.BidderName.APPNEXUS
 import static org.prebid.server.functional.model.bidder.BidderName.GENERIC
 import static org.prebid.server.functional.model.bidder.BidderName.ALIAS
@@ -32,7 +30,6 @@ import static org.prebid.server.functional.model.bidder.CompressionType.NONE
 import static org.prebid.server.functional.model.request.auction.DistributionChannel.APP
 import static org.prebid.server.functional.model.request.auction.DistributionChannel.SITE
 import static org.prebid.server.functional.model.response.auction.ErrorType.PREBID
-import static org.prebid.server.functional.model.response.auction.ImpRejectionReason.NO_BID
 import static org.prebid.server.functional.model.response.auction.MediaType.NATIVE
 import static org.prebid.server.functional.util.HttpUtil.CONTENT_ENCODING_HEADER
 import static org.prebid.server.functional.util.privacy.CcpaConsent.Signal.ENFORCED
@@ -632,27 +629,5 @@ class BidderParamsSpec extends BaseSpec {
         then: "Bidder request should contain header Content-Encoding = gzip"
         assert response.ext?.debug?.httpcalls?.get(ALIAS.value)?.requestHeaders?.first()
                        ?.get(CONTENT_ENCODING_HEADER)?.first() == compressionType
-    }
-
-    def "PBS should emit error when return all bid status fag true and bidder return error in response"() {
-        given: "Default BidRequest with returnAllBidStatus"
-        def bidRequest = BidRequest.defaultBidRequest.tap {
-            ext.prebid = new Prebid(returnAllBidStatus: true)
-        }
-
-        and: "Default basic bid with adm"
-        def bidResponse = BidResponse.getDefaultBidResponse(bidRequest)
-
-        and: "Set bidder response"
-        bidder.setResponse(bidRequest.id, BAD_REQUEST_400, bidResponse)
-
-        when: "PBS processes auction request"
-        def response = defaultPbsService.sendAuctionRequest(bidRequest)
-
-        then: "PBS response should contain seatNonBid"
-        def seatNonBid = response.ext.prebid.seatNonBid[0]
-        assert seatNonBid.seat == GENERIC.value
-        assert seatNonBid.nonBid[0].impId == bidRequest.imp[0].id
-        assert seatNonBid.nonBid[0].statusCode == NO_BID
     }
 }
